+++ semcore/dropdown-menu/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [4.39.0] - 2024-09-12
+## [4.40.0] - 2024-09-24
 
 ### Changed
 
 - Refactored to `AbstractDropdown` and `Dropdown.Item`.
+
+## [4.39.0] - 2024-09-20
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/dropdown` [4.36.4 ~> 4.37.0], `@semcore/utils` [4.35.0 ~> 4.36.0], `@semcore/core` [2.30.0 ~> 2.31.0]).
 
 ## [4.38.0] - 2024-09-06