+++ semcore/dropdown/CHANGELOG.md
@@ -2,13 +2,19 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [4.37.0] - 2024-09-12
+## [4.38.0] - 2024-09-24
 
 ### Added
 
 - `AbstractDropdown` internal class. Need it to create DropdownMenu and Select.
 - `Dropdown.Item` with styles for DropdownMenu.Item and Select.Item.
 - `Dropdown.Group` with styles for wrap DropdownMenu.Item and Select.Item into groups.
+
+## [4.37.0] - 2024-09-20
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/utils` [4.35.0 ~> 4.36.0], `@semcore/core` [2.30.0 ~> 2.31.0]).
 
 ## [4.36.4] - 2024-08-28